move image nodes into sub-folders
